Eliminate copy-pasted RPC service implementations
Replace ~7600 lines of near-identical XNV/XMR/WOW daemon and wallet
service classes with DaemonServiceBaseXmr and WalletServiceBaseXmr
abstract base classes. Leaf classes are reduced to property-only
overrides for coin prefix, block time, and atomic unit conversion.
WalletServiceXNV overrides the three wallet methods where Nerva's RPC
differs from XMR/WOW.

Remove now-unused Common* helper files and reorganize into
Rpc/Daemon/Coins/ and Rpc/Wallet/Coins/ subfolders.
