+++ NervaOneWalletMiner/Rpc/Daemon/Coins/XmrBased/DaemonServiceXMR.cs
@@ -1,0 +1,9 @@
+namespace NervaOneWalletMiner.Rpc.Daemon
+{
+    // Monero implementation as of 5/10/24: https://github.com/monero-project/monero
+    public class DaemonServiceXMR : DaemonServiceBaseXMR
+    {
+        protected override string CoinPrefix => "XMR";
+        protected override double BlockSeconds => 120.0;
+    }
+}
