+++ NervaOneWalletMiner/Rpc/Wallet/Coins/XmrBased/WalletServiceXMR.cs
@@ -1,0 +1,8 @@
+namespace NervaOneWalletMiner.Rpc.Wallet
+{
+    // Monero implementation as of 5/13/24: https://github.com/monero-project/monero
+    public class WalletServiceXMR : WalletServiceBaseXMR
+    {
+        protected override string CoinPrefix => "XMR";
+    }
+}
